+++ CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- [#221](https://github.com/os2display/display-admin-client/pull/221)
+  - Fixed missing pagination button in drag-and-drop-table.
 - [#214](https://github.com/os2display/display-admin-client/pull/214)
-  Use OIDC Code authorization flow and remove React StrictMode.
+  - Use OIDC Code authorization flow and remove React StrictMode.
 
 ## [1.4.0] - 2023-09-14